requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.5.1+cpu
torchvision==0.6.1+cpu
click==7.1.2
Flask==1.1.2
flask-cors==3.0.9
future==0.18.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.1
Pillow==7.2.0
Werkzeug==1.0.1
